dedup: extract shared MarkdownScan line classifier
IsTableLine/IsCodeFence/IsSeparatorLine/IsSeparatorCell were verbatim in both
CountOutput and MarkdownTableRowLimiter, and MarkdownSectionOrderer had its own
inline fence tracker. Move them into one internal MarkdownScan helper so fence/table
detection can't drift between the passes that post-process Markout's rendered output.
